lib/journal-club-core/document-generators/requirements.txt
python-docx>=1.1.0
reportlab>=4.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
